cms_appliction/requirement.txt
django
django-admin
djangorestframework
